Updated sorting for time so time is converted to integer when compared
